% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryfunmod1.R
\name{comparison_estimtf}
\alias{comparison_estimtf}
\title{comparison_estimtf function}
\usage{
comparison_estimtf(x, xdist, fixparam, initparam, lower, upper, method)
}
\arguments{
\item{method}{}
}
\value{

}
\description{
Function to compare TensorFlow parameter estimations with estimations from other R functions
}
