% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryfunmod1.R
\name{disableager_estimtf}
\alias{disableager_estimtf}
\title{disableager_estimtf function}
\usage{
disableager_estimtf(
  x,
  dist,
  fixparam,
  linkfun,
  initparam,
  opt,
  hyperparameters,
  maxiter,
  tolerance,
  np
)
}
\arguments{
\item{np}{}
}
\value{

}
\description{
Function to estimate distributional parameters disabling the TensorFlow eager execution mode
}
