% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_estimtf.R
\name{dist_estimtf}
\alias{dist_estimtf}
\title{dist_estimtf function}
\usage{
dist_estimtf(
  x,
  xdist = "Normal",
  fixparam = NULL,
  initparam = NULL,
  optimizer,
  hyperparameters = NULL,
  maxiter = 1000,
  tolerance = NULL,
  eager = TRUE,
  comparison = FALSE,
  lower = NULL,
  upper = NULL,
  method = "nlminb"
)
}
\arguments{
\item{x}{a vector with data}

\item{xdist}{a character indicating the name of the distribution of interest. The default value is \code{'Normal'}}

\item{fixparam}{a list of the fixed parameters of the distribution of interest. The list must contain the parameters values and names}

\item{initparam}{a list with initial values of the parameters to be estimated. The list must contain the parameters values and names}

\item{optimizer}{a character indicating the name of the TensorFlow optimizer to be used in the estimation process The default value is \code{'Adam'}}

\item{hyperparameters}{a list with the hyperparameters values of the TensorFlow optimizer}

\item{maxiter}{a positive integer indicating the maximum number of iterations for the optimization algorithm}

\item{eager}{logical. If TRUE}

\item{comparison}{logical. If TRUE the}

\item{lower}{a numeric vector with lower bounds, with the same lenght of argument `initparam`}

\item{upper}{a numeric vector with upper bounds, with the same lenght of argument `initparam`}

\item{method}{a character with the name of the optimization routine. \code{nlminb}, \code{optim}, \code{DEoptim} are available}
}
\value{
The output from
}
\description{
Function to estimate distributional parameters using TensorFlow
}
