% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_estimtf2.R
\name{dist_estimtf2}
\alias{dist_estimtf2}
\title{dist_estimtf2 function}
\usage{
dist_estimtf2(
  x,
  xdist = "Normal",
  fixparam = NULL,
  initparam = NULL,
  optimizer = "AdamOptimizer",
  hyperparameters = NULL,
  maxiter = 10000
)
}
\arguments{
\item{x}{a vector containing the data to be fitted.}

\item{xdist}{a character indicating the name of the distribution of interest or the probability density function (fdp) of the distribution. The default value is \code{'Normal'}.
The available distributions are: \code{Normal}, \code{Poisson}, \code{Binomial}, \code{Weibull}, \code{Exponential}, \code{LogNormal}, \code{Beta} and \code{Gamma}.}

\item{fixparam}{a list of the fixed parameters of the distribution of interest. The list must contain the parameters values and names. If the fdp is provided, the values for the fixed parameters must be
included in the function.}

\item{initparam}{a list with initial values of the parameters to be estimated. The list must contain the parameters values and names.}

\item{optimizer}{a character indicating the name of the TensorFlow optimizer to be used in the estimation process The default value is \code{'AdamOptimizer'}. The available optimizers are:
\code{"AdadeltaOptimizer"}, \code{"AdagradDAOptimizer"}, \code{"AdagradOptimizer"}, \code{"AdamOptimizer"}, \code{"GradientDescentOptimizer"},
\code{"MomentumOptimizer"} and \code{"RMSPropOptimizer"}.}

\item{hyperparameters}{a list with the hyperparameters values of the TensorFlow optimizer. (See URL for details of hyperparameters.)}

\item{maxiter}{a positive integer indicating the maximum number of iterations for the optimization algorithm.}
}
\value{
This function returns the estimates and standard errors of parameters from statistical distributions as well as
some information of the optimization process like the number of iterations needed for convergence.
}
\description{
Function to compute the Maximum Likelihood Estimators of distributional parameters using TensorFlow.
}
\details{
\code{dist_estimtf2} computes the log-likelihood function of the distribution specified in
\code{xdist} and finds the distributional parameters that maximizes it using TensorFlow.
}
\note{
The \code{summary, print} functions can be used with a \code{dist_estimtf2} object.
}
\examples{
#-------------------------------------------------------------
# Estimation of both normal distrubution parameters
x <- rnorm(n = 1000, mean = 10, sd = 3)

estimation_1 <- dist_estimtf2(x, xdist = "Normal", optimizer = "AdamOptimizer",
                           hyperparameters = list(learning_rate = 0.1))

summary(estimation_1)

#-------------------------------------------------------------
# Estimation of parameters from Instantaneous Failures distribution
pdf <- function(X, lambda) { (1 / ((lambda ^ 2) * (lambda - 1))) * (lambda^2 + X - 2*lambda) * exp(-X/lambda) }
x <-  c(3.4, 0.0, 0.0, 15.8, 232.8, 8.8, 123.2, 47, 154, 103.2, 89.8,  12.2)

estimation_2 <- dist_estimtf2(x = x, xdist = pdf, initparam = list(lambda = rnorm(1, 2, 1)), optimizer = "AdamOptimizer",
                              hyperparameters = list(learning_rate = 0.1),
                              maxiter = 10000)
summary(estimation_2)

}
\author{
Sara Garces Cespedes
}
