% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaryfunmod1.R
\name{eager_estimtf}
\alias{eager_estimtf}
\title{eager_estimtf function}
\usage{
eager_estimtf(
  x,
  dist,
  fixparam,
  linkfun,
  initparam,
  opt,
  hyperparameters,
  maxiter,
  tolerance,
  np
)
}
\arguments{
\item{np}{}
}
\value{

}
\description{
Function to estimate distributional parameters in TensorFlow eager execution mode
}
