% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimglm.R
\name{estimglm}
\alias{estimglm}
\title{estimglm function}
\usage{
estimglm(
  formula,
  family = "Normal",
  link_function = "identity",
  data = NULL,
  initcoeff = NULL,
  optimizer = "AdamOptimizer",
  hyperparameters = NULL,
  maxiter = 10000
)
}
\arguments{
\item{formula}{an object of class "formula" that specifies the model to be fitted.}

\item{family}{a description of the error distribution to be used in the model.}

\item{link_function}{a string that specifies the model link function.}

\item{data}{an optional data frame containing the variables in the model. If these variables are
not found in \code{data}, they ara taken from the environment from which \code{reg_estimtf} is called.}

\item{initcoeff}{Optional a list with the initial values for model coefficients. Default value: Zeros.}

\item{optimizer}{a character indicating the name of the TensorFlow optimizer to be used in the estimation process The default value is \code{'AdamOptimizer'}. The available optimizers are:
\code{"AdadeltaOptimizer"}, \code{"AdagradDAOptimizer"}, \code{"AdagradOptimizer"}, \code{"AdamOptimizer"}, \code{"GradientDescentOptimizer"},
\code{"MomentumOptimizer"} and \code{"RMSPropOptimizer"}.}

\item{hyperparameters}{a list with the hyperparameters values of the TensorFlow optimizer. FALTA DETALLES}

\item{maxiter}{a positive integer indicating the maximum number of iterations for the optimization algorithm. The default value is \code{10000}.}
}
\value{
This function returns the estimates for coefficients of Generalized Linear Models. ESTA MAL
}
\description{
Function to fit Generalized Linear Models using TensorFlow.
}
\details{
\code{estimglm} computes the log-likelihood function depending on the provided family. Then, it finds the regression coefficients
that maximizes it using TensorFlow. ESTA MAL
}
\examples{
#-------------------------------------------------------------
# Fit simple linear regression model (cars dataset)
data <- data.frame(dist = cars$dist, speed = cars$speed)
formula <- dist ~ speed
estimation_1 <- estimglm(formula, family = "Normal", link_function = "identity",
                         data = data, initcoeff = list(speed = 1),
                         optimizer = "AdamOptimizer", hyperparameters = list(learning_rate = 0.1),
                         maxiter = 10000)

summary(estimation_1)

}
\author{
Sara Garces Cespedes
}
