% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_estimtf.R
\name{glm_estimtf}
\alias{glm_estimtf}
\title{glm_estimtf function}
\usage{
glm_estimtf(
  formula,
  family = "Normal",
  link_function = "identity",
  data = NULL,
  initcoeff = NULL,
  maxiter = 10000
)
}
\arguments{
\item{formula}{an object of class "formula" that specifies the model to be fitted.}

\item{family}{a description of the error distribution to be used in the model.}

\item{link_function}{a string that specifies the model link function.}

\item{data}{an optional data frame containing the variables in the model. If these variables are
not found in \code{data}, they ara taken from the environment from which \code{reg_estimtf} is called.}

\item{initcoeff}{Optional a list with the initial values for model coefficients. Default value: Zeros.}

\item{maxiter}{a positive integer indicating the maximum number of iterations for the optimization algorithm. The default value is \code{10000}.}
}
\value{
This function returns the estimates for parameters of Generalized Linear Models. These parameters are
estimated using the Fisher scoring method.
}
\description{
Function to compute the Maximum Likelihood Estimators of regression parameters from Generalized
Linear Models using TensorFlow.
}
\details{
\code{glm_estimtf} uses the Fisher scoring method to determine the maximum of the likelihood function.
}
\examples{

ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
weight <- c(ctl, trt)

family <- "Normal"
link_function <- "identity"
formula <- weight ~ group
data <- data.frame(weight = weight, group = group)
estimation <- glm_estimtf(formula = formula, family = family,
                          link_function = link_function, data = data)

print(estimation)

}
\author{
Sara Garces Cespedes
}
