% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mleregtf.R
\name{mleregtf}
\alias{mleregtf}
\title{mleregtf function}
\usage{
mleregtf(
  ydist = y ~ Normal,
  formulas,
  data,
  available_distribution = TRUE,
  fixparam = NULL,
  initparam = NULL,
  link_function = NULL,
  optimizer = "AdamOptimizer",
  hyperparameters = NULL,
  maxiter = 10000
)
}
\arguments{
\item{ydist}{an object of class "formula" that specifies the distribution of the response variable.
The available distributions are: \code{Normal}, \code{Poisson}, \code{Binomial}, \code{Weibull}, \code{Exponential}, \code{LogNormal},
\code{Beta} and \code{Gamma}. If you want to estimate parameters from a distribution not included in the package, you must provide the
name of an \code{R} function that contains its probability mass/density function.}

\item{formulas}{a list containing objects of class "formula". Each element of the list represents the
linear predictor for each of the parameters of the regression model. The linear predictor is specified with
the name of the parameter followed by \code{.fo} and it must contain an \code{~}. The terms on the right side
separated by \code{+}.}

\item{data}{a data frame containing the response variable and the covariates.}

\item{available_distribution}{logical. If TRUE, the distribution of the response variable is one of the available distributions in the package.}

\item{fixparam}{a list containing the fixed parameters of the model. The parameters values and names must be specified in the list.}

\item{initparam}{a list with initial values of the regression coefficients to be estimated. The list must contain the regression coefficients values and names.
If you want to use the same initial values for all regression coefficients associated with a specific parameter, you can specify the
name of the parameter and the value. If NULL the default initial value is zero.}

\item{link_function}{a list with names of parameters to be linked and the corresponding link function name. The available link functions are:
\code{log}, \code{logit}, \code{inverse} and \code{identity}.}

\item{optimizer}{a character indicating the name of the TensorFlow optimizer to be used in the estimation process. The default value is \code{'AdamOptimizer'}. The available optimizers are:
\code{"AdadeltaOptimizer"}, \code{"AdagradDAOptimizer"}, \code{"AdagradOptimizer"}, \code{"AdamOptimizer"}, \code{"GradientDescentOptimizer"},
\code{"MomentumOptimizer"} and \code{"RMSPropOptimizer"}.}

\item{hyperparameters}{a list with the hyperparameters values of the selected TensorFlow optimizer. If the hyperparameters are not specified, their default values
will be used in the oprimization process (See URL for details of hyperparameters.)}

\item{maxiter}{a positive integer indicating the maximum number of iterations for the optimization algorithm. The default value is \code{10000}.}
}
\value{
This function returns the estimates, standard errors, Z-score and p-values of significance tests
of the regression model coefficients as well as some information of the optimization process like the number of
iterations needed for convergence.
}
\description{
Function to compute the Maximum Likelihood Estimators of regression parameters with linear predictors using TensorFlow.
}
\details{
\code{mleregtf} computes the log-likelihood function based on the distribution specified in
\code{ydist} and linear predictors specified in \code{formulas}. Then, it finds the values of the regression coefficients
that maximizes this function using the TensorFlow opimizer specified in \code{optimizer}.
}
\note{
The \code{summary, print} functions can be used with a \code{mleregtf} object.
}
\examples{
#-------------------------------------------------------------
# Estimation of parameters of a Poisson regression model
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
data <- data.frame(treatment, outcome, counts)

formulas <- list(lambda.fo = ~ outcome + treatment)
estimation_1 <- mleregtf(ydist =  counts ~ Poisson,
                            formulas = formulas,
                            data = data,
                            initparam = list(lambda = 1.0),
                            optimizer = "AdamOptimizer",
                            link_function = list(lambda = "log"),
                            hyperparameters = list(learning_rate = 0.1))
summary(estimation_1)

#-------------------------------------------------------------
# Estimation of parameters of a linear regression model with one fixed parameter
x <- runif(n = 1000, -3, 3)
y <- rnorm(n = 1000, mean = 5 - 2 * x, sd = 3)
data <- data.frame(y = y, x = x)

formulas <- list(mean.fo = ~ x)
initparam <- list(mean = list(Intercept = 1.0, x = 0.0))

estimation_2 <- mleregtf(ydist = y ~ Normal,
                            formulas = formulas,
                            data = data,
                            fixparam = list(sd = 3),
                            initparam = initparam,
                            optimizer = "AdamOptimizer",
                            hyperparameters = list(learning_rate = 0.1))

summary(estimation_2)

#-------------------------------------------------------------
# Estimation of parameters from Instantaneous Failures distribution
pdf <- function(y, lambda) { (1 / ((lambda ^ 2) * (lambda - 1))) *
                             (lambda^2 + y - 2*lambda) * exp(-y/lambda) }
y <-  c(3.4, 0.0, 0.0, 15.8, 232.8, 8.8, 123.2, 47, 154, 103.2, 89.8,  12.2)
data <- data.frame(y)

estimation_3 <- mleregtf(ydist = y ~ pdf,
                             formulas = list(lambda.fo = ~1),
                             data = data,
                             initparam = list(lambda = rnorm(1, 5, 1)),
                             available_distribution = FALSE,
                             optimizer = "AdamOptimizer",
                             hyperparameters = list(learning_rate = 0.1),
                             maxiter = 10000)

summary(estimation_3)


}
\author{
Sara Garces Cespedes
}
