% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.R
\name{plot_loss}
\alias{plot_loss}
\title{plot_loss function}
\usage{
plot_loss(object, ...)
}
\arguments{
\item{object}{an object of class \code{MLEtf} for which a plot with loss values is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Function to display a graph that contains the loss value in each iteration of
the estimation process using \code{\link{dist_estimtf}} function or using \code{\link{reg_estimtf}} function.
}
\details{
\code{plot_loss.MLEtf} function displays a graph with the loss values in each iteration of the
estimation process for distributional or regression parameters.
}
\examples{
#-------------------------------------------------------------
# Estimation of both normal distrubution parameters
x <- rnorm(n = 1000, mean = 10, sd = 3)

plot_loss(dist_estimtf(x, xdist = "Normal", optimizer = "AdamOptimizer",
                           hyperparameters = list(learning_rate = 0.1)))

}
\author{
Sara Garces Cespedes
}
