% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_function.R
\name{predict.MLEtf}
\alias{predict.MLEtf}
\title{predict.MLEtf function}
\usage{
\method{predict}{MLEtf}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{MLEtf} for which a summary is desired.}

\item{newdata}{an optional data frame in which to look variables with which to predict. If \code{newdata} is
missing or \code{NULL}, the fitted values are used.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
The output from
}
\description{
Function to produce result summaries of the estimates of parameters from statistical
distributions using \code{\link{dist_estimtf2}} or parameters from regression models using
\code{\link{reg_estimtf}}.
}
\details{
\code{predict.MLEtf} COMPLETAR
}
\author{
Sara Garces Cespedes
}
