% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.R
\name{print.MLEtf}
\alias{print.MLEtf}
\title{print.MLEtf function}
\usage{
\method{print}{MLEtf}(x, ...)
}
\arguments{
\item{x}{an object of class \code{MLEtf} for which the estimates are desired.}

\item{...}{additional arguments affecting the estimates produced.}
}
\description{
Function to display the estimates of parameters from probability
distributions using the \code{\link{mle_tf}} function or parameters from regression models using
the \code{\link{mlereg_tf}} function.
}
\details{
\code{print.MLEtf} function displays the estimates of parameters from probability distributions
and regression models.
}
\examples{
#---------------------------------------------------------------
# Estimation of both normal distrubution parameters

# Generate a sample from the normal distribution
x <- rnorm(n = 1000, mean = 10, sd = 3)

# Use the print function
print(mle_tf(x,
             xdist = "Normal",
             initparam = list(mean = 1.0, sd = 1.0),
             optimizer = "AdamOptimizer",
             hyperparameters = list(learning_rate = 0.1)))

}
\author{
Sara Garcés Céspedes \email{sgarcesc@unal.edu.co}
}
