% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_estimtf.R
\name{reg_estimtf}
\alias{reg_estimtf}
\title{reg_estimtf function}
\usage{
reg_estimtf(
  ydist = y ~ Normal,
  formulas,
  data = NULL,
  fixparam = NULL,
  initparam = NULL,
  link_function = NULL,
  optimizer = "AdamOptimizer",
  hyperparameters = NULL,
  maxiter = 10000,
  eager = TRUE
)
}
\arguments{
\item{ydist}{an object of class "formula" that specifies the distribution of the response variable. FALTA}

\item{formulas}{a list containing objects of class "formula". Each element of the list specifies the
linear predictor for each of the parameters of the distribution of interest. FALTA}

\item{data}{an optional data frame containing the variables in the model. If these variables are
not found in \code{data}, they ara taken from the environment from which \code{reg_estimtf} is called.}

\item{fixparam}{a list of the fixed parameters of the distribution of interest. The list must contain the parameters values and names.}

\item{initparam}{a list with initial values of the regression parameters to be estimated. The list must contain the regression parameters values and names.
If you want to use the same initial values for all regression parameters associated with a distributional parameter, you can write the
name of the distributional parameter and the value.}

\item{link_function}{a list with names of parameters to be linked and the corresponding link function name. The link functions available are:
\code{log} and \code{logit}.}

\item{optimizer}{a character indicating the name of the TensorFlow optimizer to be used in the estimation process The default value is \code{'AdamOptimizer'}.}

\item{hyperparameters}{a list with the hyperparameters values of the TensorFlow optimizer. FALTA DETALLES}

\item{maxiter}{a positive integer indicating the maximum number of iterations for the optimization algorithm. The default value is \code{10000}.}

\item{eager}{If \code{TRUE}, the estimation process is performed in the eager execution environment. The default value is \code{TRUE}.}
}
\value{
This function returns the estimates, standard errors, z scores and p-values of significance tests
for the regression parameters as well as some information of the optimization process like the number of
iterations needed for convergence.
}
\description{
Function to compute the Maximum Likelihood Estimators of regression parameters using TensorFlow
}
\details{
\code{reg_estimtf} computes the log-likelihood function of the distribution specified in
\code{ydist} with linear predictors specified in \code{formulas}. Then, it finds the regression parameters
that maximizes it using TensorFlow.

#' @examples
#-------------------------------------------------------------
# Estimation of parameters of a linear regression model
n <- 1000
x <- runif(n = 1000, 0, 10)
x1 <- runif(n = n, 0, 6)
y <- rnorm(n = n, mean = -2 + 3 * x + 9 * x1, sd = exp(3 + 3 * x))
data <- data.frame(y = y, x = x, x1 = x1)

formulas <- list(loc.fo = ~ x + x1, scale.fo = ~ x)

estimation_1 <- reg_estimtf(ydist = "Normal", formulas = formulas, data = data,
                            initparam <- list(loc = 1.0, scale = 1.0),
                            link_function <- list(scale = "log"), optimizer = "AdamOptimizer",
                            hyperparameters = list(learning_rate = 0.1))

summary(estimation_1)

#-------------------------------------------------------------
# Estimation of parameters of a linear regression model with one fixed parameter
x <- runif(n = 1000, -3, 3)
y <- rnorm(n = 1000, mean = 5 - 2 * x, sigma = 3)
data <- data.frame(y = y, x = x)

formulas <- list(loc.fo = ~ x)
initparam <- list(loc = list(Intercept = 1.0, x = 0.0))

estimation_2 <- reg_estimtf(ydist = "Normal", formulas = formulas, data = data,
                            fixparam = list(scale = 3), initparam = initparam
                            link_function <- list(scale = "log"), optimizer = "AdamOptimizer",
                            hyperparameters = list(learning_rate = 0.1))

summary(estimation_2)
}
\author{
Sara Garces Cespedes
}
