% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.R
\name{summary.MLEtf}
\alias{summary.MLEtf}
\title{summary.MLEtf function}
\usage{
\method{summary}{MLEtf}(object, ...)
}
\arguments{
\item{object}{an object of class \code{MLEtf} for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
The output from
}
\description{
Function to produce result summaries of the estimates of parameters from statistical
distributions using \code{\link{dist_estimtf2}} or parameters from regression models using
\code{\link{mleregtf}}.
}
\details{
\code{summary.MLEtf} function displays estimates and standard errors of parameters from statistical
distributions and regression models. Also, this function computes and displays the Z-score and p-values of significance
tests for these parameters.
}
\examples{
#-------------------------------------------------------------
# Estimation of both normal distrubution parameters
x <- rnorm(n = 1000, mean = 10, sd = 3)

summary(dist_estimtf2(x, xdist = "Normal",
                      optimizer = "AdamOptimizer",
                      initparam = list(mean = 1.0, sd = 1.0),
                      hyperparameters = list(learning_rate = 0.1)))

}
\author{
Sara Garces Cespedes
}
