% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_estimtf.R
\name{dist_estimtf}
\alias{dist_estimtf}
\title{dist_estimtf function}
\usage{
dist_estimtf(
  x,
  xdist = "Normal",
  fixparam = NULL,
  initparam = NULL,
  optimizer = "AdamOptimizer",
  hyperparameters = NULL,
  maxiter = 10000
)
}
\arguments{
\item{x}{a vector containing the data to be fitted.}

\item{xdist}{a character indicating the name of the distribution of interest. The default value is \code{'Normal'}.}

\item{fixparam}{a list of the fixed parameters of the distribution of interest. The list must contain the parameters values and names.}

\item{initparam}{a list with initial values of the parameters to be estimated. The list must contain the parameters values and names.}

\item{optimizer}{a character indicating the name of the TensorFlow optimizer to be used in the estimation process The default value is \code{'AdamOptimizer'}. The available optimizers are:
\code{"AdadeltaOptimizer"}, \code{"AdagradDAOptimizer"}, \code{"AdagradOptimizer"}, \code{"AdamOptimizer"}, \code{"GradientDescentOptimizer"},
\code{"MomentumOptimizer"} and \code{"RMSPropOptimizer"}.}

\item{hyperparameters}{a list with the hyperparameters values of the TensorFlow optimizer. FALTA DETALLES}

\item{maxiter}{a positive integer indicating the maximum number of iterations for the optimization algorithm.}
}
\value{
This function returns the estimates and standard errors of parameters from statistical distributions as well as
some information of the optimization process like the number of iterations needed for convergence.
}
\description{
Function to compute the Maximum Likelihood Estimators of distributional parameters using TensorFlow.
}
\details{
\code{dist_estimtf} computes the log-likelihood function of the distribution specified in
\code{xdist} and finds the distributional parameters that maximizes it using TensorFlow.
}
\examples{
#-------------------------------------------------------------
# Estimation of both normal distrubution parameters
x <- rnorm(n = 1000, mean = 10, sd = 3)

estimation_1 <- dist_estimtf(x, xdist = "Normal", optimizer = "AdamOptimizer",
                           hyperparameters = list(learning_rate = 0.1))

summary(estimation_1)

#-------------------------------------------------------------
# Estimation with one fixed parameter
x <- rnorm(n = 1000, mean = 10, sd = 3)

estimation_2 <- dist_estimtf(x, xdist = "Normal", optimizer = "AdamOptimizer",
                           hyperparameters = list(learning_rate = 0.1),
                           fixparam = list(loc = 10))

summary(estimation_2)


}
\author{
Sara Garces Cespedes
}
