% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.R
\name{print.MLEtf}
\alias{print.MLEtf}
\title{print.MLEtf function}
\usage{
\method{print}{MLEtf}(x, ...)
}
\arguments{
\item{x}{an object of class \code{MLEtf} for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Function to display the estimates of parameters from statistical
distributions using \code{\link{dist_estimtf2}} or parameters from regression models using
\code{\link{reg_estimtf}}.
}
\details{
\code{print.MLEtf} function displays estimates of parameters from statistical distributions
and regression models.
}
\examples{
#-------------------------------------------------------------
# Estimation of both normal distrubution parameters
x <- rnorm(n = 1000, mean = 10, sd = 3)

print(dist_estimtf2(x, xdist = "Normal",
                    initparam = list(mean = 1.0, sd = 1.0),
                    optimizer = "AdamOptimizer",
                    hyperparameters = list(learning_rate = 0.1)))

}
\author{
Sara Garces Cespedes
}
