% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_estimtf.R
\name{reg_estimtf}
\alias{reg_estimtf}
\title{reg_estimtf function}
\usage{
reg_estimtf(
  ydist = y ~ Normal,
  formulas,
  data = NULL,
  fixparam = NULL,
  initparam = NULL,
  link_function = NULL,
  optimizer = "AdamOptimizer",
  hyperparameters = NULL,
  maxiter = 10000
)
}
\arguments{
\item{ydist}{an object of class "formula" that specifies the distribution of the response variable. The available distributions are:}

\item{formulas}{a list containing objects of class "formula". Each element of the list represents the
linear predictor for each of the parameters of the regression model. The linear predictor is specified with
the name of the parameter followed by \code{.fo} and it must contain an \code{~} and the terms on the right side
separated by \code{+}.}

\item{data}{a data frame containing the variables in the regression model.}

\item{fixparam}{a list containing the fixed parameters of the model. The parameters values and names must be specified in the list.}

\item{initparam}{a list with initial values of the regression coefficients to be estimated. The list must contain the regression coefficients values and names.
If you want to use the same initial values for all regression coefficients associated with a specific parameter, you can specify the
name of the parameter and the value.}

\item{link_function}{a list with names of parameters to be linked and the corresponding link function name. The available link functions are:
\code{log}, \code{logit}, \code{inverse} and \code{identity}.}

\item{optimizer}{a character indicating the name of the TensorFlow optimizer to be used in the estimation process. The default value is \code{'AdamOptimizer'}. The available optimizers are:
\code{"AdadeltaOptimizer"}, \code{"AdagradDAOptimizer"}, \code{"AdagradOptimizer"}, \code{"AdamOptimizer"}, \code{"GradientDescentOptimizer"},
\code{"MomentumOptimizer"} and \code{"RMSPropOptimizer"}.}

\item{hyperparameters}{a list with the hyperparameters values of the TensorFlow optimizer. (See URL for details of hyperparameters.)}

\item{maxiter}{a positive integer indicating the maximum number of iterations for the optimization algorithm. The default value is \code{10000}.}
}
\value{
This function returns the estimates, standard errors, t values and p-values of significance tests
for the regression model coefficients as well as some information of the optimization process like the number of
iterations needed for convergence.
}
\description{
Function to compute the Maximum Likelihood Estimators of regression parameters with linear predictors using TensorFlow.
}
\details{
\code{reg_estimtf} computes the log-likelihood function based on the distribution specified in
\code{ydist} and linear predictors specified in \code{formulas}. Then, it finds the regression coefficients
that maximizes it using TensorFlow.
}
\note{
The \code{summary, print} functions can be used with a \code{reg_estimtf} object.
}
\examples{
#-------------------------------------------------------------
# Estimation of parameters of a Poisson regression model
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
data <- data.frame(treatment, outcome, counts)

formulas <- list(rate.fo = ~ outcome + treatment)
estimation_1 <- reg_estimtf(ydist =  counts ~ Poisson, formulas = formulas, data = data,
                            initparam = list(rate = 1.0), optimizer = "AdamOptimizer",
                            link_function = list(rate = "log"),
                            hyperparameters = list(learning_rate = 0.1))
summary(estimation_1)

#-------------------------------------------------------------
# Estimation of parameters of a linear regression model with one fixed parameter
x <- runif(n = 1000, -3, 3)
y <- rnorm(n = 1000, mean = 5 - 2 * x, sd = 3)
data <- data.frame(y = y, x = x)

formulas <- list(loc.fo = ~ x)
initparam <- list(loc = list(Intercept = 1.0, x = 0.0))

estimation_2 <- reg_estimtf(ydist = y ~ Normal, formulas = formulas, data = data,
                            fixparam = list(scale = 3), initparam = initparam,
                            optimizer = "AdamOptimizer",
                            hyperparameters = list(learning_rate = 0.1))

summary(estimation_2)

#-------------------------------------------------------------
# Estimation of parameters using cars dataset
data <- data.frame(dist = cars$dist, speed = cars$speed)
formulas <- list(loc.fo = ~ speed, scale.fo = ~ 1)

estimation_3 <- reg_estimtf(ydist = dist ~ Normal, formulas = formulas, data = data,
                            initparam = list(loc = 1.0, scale = 1.0), optimizer = "AdamOptimizer",
                            hyperparameters = list(learning_rate = 0.1))

summary(estimation_3)

}
\author{
Sara Garces Cespedes
}
