% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary function.R
\name{summary.MLEtf}
\alias{summary.MLEtf}
\title{summary.MLEtf function}
\usage{
\method{summary}{MLEtf}(object, ...)
}
\arguments{
\item{object}{an object of class \code{MLEtf} for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
The output from
}
\description{
Function to produce result summaries of the estimates of distributional
parameters using \code{\link{dist_estimtf}} or regression parameters using \code{\link{reg_estimtf}}.
}
\details{
\code{summary.MLEtf} function displays estimates and standard errors of distributional and
regression parameters. Also, for regression parameters, this function computes and displays the z-score
and p-values of significance tests for these parameters.
}
\examples{
#-------------------------------------------------------------
# Estimation of both normal distrubution parameters
x <- rnorm(n = 1000, mean = 10, sigma = 3)

summary(dist_estimtf(x, xdist = "Normal", optimizer = "AdamOptimizer",
                           hyperparameters = list(learning_rate = 0.1)))

}
\author{
Sara Garces Cespedes
}
