% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_function.R
\name{summary.MLEtf}
\alias{summary.MLEtf}
\title{summary.MLEtf function}
\usage{
\method{summary}{MLEtf}(object, ...)
}
\arguments{
\item{object}{an object of class \code{MLEtf} for which a summary is desired.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Function to produce result summaries of the estimates of parameters from probability
distributions using the \code{\link{mle_tf}} function or parameters from regression models using
the \code{\link{mlereg_tf}} function.
}
\details{
\code{summary.MLEtf} function displays estimates and standard errors of parameters from statistical
distributions and regression models. Also, this function computes and displays the Z-score and p-values of significance
tests for these parameters.
}
\examples{
#---------------------------------------------------------------
# Estimation of both normal distrubution parameters

# Generate a sample from the normal distribution
x <- rnorm(n = 1000, mean = 10, sd = 3)

# Use the summary function
summary(mle_tf(x, xdist = "Normal",
               optimizer = "AdamOptimizer",
               initparam = list(mean = 1.0, sd = 1.0),
               hyperparameters = list(learning_rate = 0.1)))

}
\author{
Sara Garcés Céspedes \email{sgarcesc@unal.edu.co}
}
